chore(tui): split render cells into sub-modules and add /review command
- Extract ActiveTurnCell (~486 lines) into cells/active_turn.rs
- Extract CommittedTurnCell (~104 lines) into cells/committed_turn.rs
- Cells/mod.rs reduced from 938 to 344 lines
- Move section_span and wrapped_history_line_count into render/helpers.rs
- Add /review slash command: captures git diff, composes review prompt
  for the agent (LocalCommandKind, CommandSpec, handler)
