+++ src/tui/runtime/commands.rs
@@ -28,6 +28,7 @@ pub(super) async fn execute_local_command(
         LocalCommandKind::Plan => "plan",
         LocalCommandKind::Quit => "quit",
         LocalCommandKind::Resume => "resume",
+        LocalCommandKind::Review => "review",
         LocalCommandKind::Status => "status",
     });
     match command.kind {
@@ -113,6 +114,40 @@ pub(super) async fn execute_local_command(
             }
             app.push_notice("Planning mode enabled. Read-only planning; approve to execute.");
         }
+        LocalCommandKind::Review => {
+            if app.is_busy() {
+                app.push_notice("A task is already running. Wait for it to finish.");
+            } else {
+                app.set_runtime_phase(
+                    RuntimePhase::LocalCommand,
+                    Some("preparing review prompt".into()),
+                );
+                let diff = capture_git_diff();
+                let mut prompt = String::from(
+                    "Please review the following code changes. ",
+                );
+                prompt.push_str(
+                    "Point out potential issues, suggest improvements, and highlight any concerns about correctness, performance, or maintainability.\n\n",
+                );
+                if diff.is_empty() {
+                    prompt.push_str("No local git changes found. The working tree is clean.");
+                } else if diff.lines().count() > 800 {
+                    let preview: String = diff
+                        .lines()
+                        .take(600)
+                        .map(|l| format!("{l}\n"))
+                        .collect();
+                    prompt.push_str(&format!(
+                        "```diff\n{}...\n```\n\n(Truncated; use the agent tools to inspect the full diff if needed.)",
+                        preview
+                    ));
+                } else {
+                    prompt.push_str(&format!("```diff\n{diff}\n```"));
+                }
+                app.set_input(prompt);
+                app.push_notice("Review prompt ready. Press Enter to send it to the agent.");
+            }
+        }
         LocalCommandKind::Quit => {
             app.set_runtime_phase(RuntimePhase::LocalCommand, Some("quitting".into()));
             return Ok(true);
@@ -159,4 +194,24 @@ fn handle_base_url_command(arg: Option<&str>, app: &mut TuiApp) -> anyhow::Resul
     }
     app.open_overlay(Overlay::BaseUrlEditor);
     Ok(())
+}
+
+fn capture_git_diff() -> String {
+    use std::process::Command;
+    let run = |args: &[&str]| -> Option<String> {
+        Command::new("git")
+            .args(args)
+            .output()
+            .ok()
+            .and_then(|out| String::from_utf8(out.stdout).ok())
+            .filter(|s| !s.trim().is_empty())
+    };
+    let staged = run(&["diff", "--staged"]);
+    let unstaged = run(&["diff"]);
+    match (staged, unstaged) {
+        (Some(s), Some(u)) => format!("{s}\n{u}"),
+        (Some(s), None) => s,
+        (None, Some(u)) => u,
+        (None, None) => String::new(),
+    }
 }
