+++ src/tui/state/types.rs
@@ -86,6 +86,7 @@ pub enum LocalCommandKind {
     BaseUrl,
     Login,
     Logout,
+    Review,
     Quit,
 }
